Added unit tests.
