+++ quickwit-serve/Cargo.toml
@@ -35,6 +35,7 @@ tokio-stream = "0.1.6"
 mockall = "0.10"
 assert-json-diff = "2.0.1"
 tokio = { version = "1", features = ["full"] }
+quickwit-storage = { version = "0.1.0", path = "../quickwit-storage", features=["testsuite"]}
 
 [dependencies.quickwit-cluster]
 path = '../quickwit-cluster'
